Extract engine limits policy
